refactor: migrate frontend to MVVM architecture
- Extract business logic from screens into custom hooks (ViewModels)
- Create reusable hooks: useFeed, useLibrary, useProfile, useMyApps, useStore, useMicroApp, useAppLayout
- Add AppInitializer component for app state initialization
